Gate ECDSA support behind a feature
