+++ Cargo.toml
@@ -12,8 +12,9 @@ rust-version = "1.79"
 [dependencies]
 primeorder = { version = "0.13", default-features = false }
 num-traits = { version = "0.2", default-features = false }
-ecdsa = { version = "0.16", default-features = false, features = ["signing", "verifying"] }
 sha2 = { version = "0.10", default-features = false }
+
+ecdsa = { version = "0.16", default-features = false, optional = true }
 
 [dev-dependencies]
 criterion = "0.5"
@@ -23,6 +24,7 @@ proptest = "1"
 
 [features]
 serde = ["primeorder/serde"]
+ecdsa = ["ecdsa/signing", "ecdsa/verifying"]
 
 [[bench]]
 bench = true
